+++ directives/date-picker.js
@@ -349,10 +349,10 @@ angular.module('keta.directives.DatePicker', [
 						scope.showPager : true;
 				scope.showSelectionButton =
 					angular.isDefined(scope.showSelectionButton) ?
-						scope.showSelectionButton : true;
+						scope.showSelectionButton : false;
 				scope.showTodayButton =
 					angular.isDefined(scope.showTodayButton) ?
-						scope.showTodayButton : true;
+						scope.showTodayButton : false;
 				scope.showWeekNumbers =
 					angular.isDefined(scope.showWeekNumbers) ?
 						scope.showWeekNumbers : true;